move predict to tile_predict and rename imports
